🐛 Fix exclude config behavior (#1695)

+++ packages/core/src/service/Project.ts
@@ -1,5 +1,4 @@
-import type { Ignore } from 'ignore'
-import ignore from 'ignore'
+import * as micromatch from 'micromatch'
 import type { TextDocumentContentChangeEvent } from 'vscode-languageserver-textdocument'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import type { ExternalEventEmitter, Externals, FsWatcher, IntervalId } from '../common/index.js'
@@ -158,7 +157,6 @@ export type ProjectData = Pick<
  */
 export class Project implements ExternalEventEmitter {
 	private static readonly RootSuffix = '/pack.mcmeta'
-	private static readonly GitIgnore = '.gitignore'
 
 	/** Prevent circular binding. */
 	readonly #bindingInProgressUris = new Set<string>()
@@ -183,7 +181,6 @@ export class Project implements ExternalEventEmitter {
 	}
 
 	config!: Config
-	ignore: Ignore = ignore()
 	readonly downloader: Downloader
 	readonly externals: Externals
 	readonly fs: FileService
@@ -297,11 +294,7 @@ export class Project implements ExternalEventEmitter {
 		this.logger.info(
 			`[Project#getTrackedFiles] Listed ${supportedFiles.length} supported files`,
 		)
-		const filteredFiles = this.ignore.filter(supportedFiles)
-		this.logger.info(
-			`[Project#getTrackedFiles] After ignoring, keeping ${filteredFiles.length} tracked files`,
-		)
-		return filteredFiles
+		return supportedFiles
 	}
 
 	constructor(
@@ -399,20 +392,6 @@ export class Project implements ExternalEventEmitter {
 	}
 
 	private setInitPromise(): void {
-		const loadConfig = async () => {
-			this.config = await this.#configService.load()
-			this.ignore = ignore()
-			for (const pattern of this.config.env.exclude) {
-				if (pattern === '@gitignore') {
-					const gitignore = await this.readGitignore()
-					if (gitignore) {
-						this.ignore.add(gitignore)
-					}
-				} else {
-					this.ignore.add(pattern)
-				}
-			}
-		}
 		const callIntializers = async () => {
 			const initCtx: ProjectInitializerContext = {
 				cacheRoot: this.cacheRoot,
@@ -446,29 +425,13 @@ export class Project implements ExternalEventEmitter {
 			this.symbols.buildCache()
 			__profiler.task('Load Cache')
 
-			await loadConfig()
+			this.config = await this.#configService.load()
 			__profiler.task('Load Config')
 
 			await callIntializers()
 			__profiler.task('Initialize').finalize()
 		}
 		this.#initPromise = init()
-	}
-
-	private async readGitignore() {
-		if (this.projectRoots.length === 0) {
-			return undefined
-		}
-		try {
-			const uri = this.projectRoots[0] + Project.GitIgnore
-			const contents = await this.externals.fs.readFile(uri)
-			return bufferToString(contents)
-		} catch (e) {
-			if (!this.externals.error.isKind(e, 'ENOENT')) {
-				this.logger.error(`[Project] [readGitignore]`, e)
-			}
-		}
-		return undefined
 	}
 
 	private setReadyPromise(): void {
@@ -529,15 +492,24 @@ export class Project implements ExternalEventEmitter {
 					this.#watcherReady = true
 					resolve()
 				}).on('add', (uri) => {
+					if (this.shouldExclude(uri)) {
+						return
+					}
 					this.#watchedFiles.add(uri)
 					if (this.#watcherReady) {
 						this.emit('fileCreated', { uri })
 					}
 				}).on('change', (uri) => {
+					if (this.shouldExclude(uri)) {
+						return
+					}
 					if (this.#watcherReady) {
 						this.emit('fileModified', { uri })
 					}
 				}).on('unlink', (uri) => {
+					if (this.shouldExclude(uri)) {
+						return
+					}
 					this.#watchedFiles.delete(uri)
 					if (this.#watcherReady) {
 						this.emit('fileDeleted', { uri })
@@ -903,6 +875,9 @@ export class Project implements ExternalEventEmitter {
 		if (!fileUtil.isFileUri(uri)) {
 			return // We only accept `file:` scheme for client-managed URIs.
 		}
+		if (this.shouldExclude(uri)) {
+			return
+		}
 		const doc = TextDocument.create(uri, languageID, version, content)
 		const node = this.parse(doc)
 		this.#clientManagedUris.add(uri)
@@ -926,6 +901,9 @@ export class Project implements ExternalEventEmitter {
 		this.#symbolUpToDateUris.delete(uri)
 		if (!fileUtil.isFileUri(uri)) {
 			return // We only accept `file:` scheme for client-managed URIs.
+		}
+		if (this.shouldExclude(uri)) {
+			return
 		}
 		const doc = this.#clientManagedDocAndNodes.get(uri)?.doc
 		if (!doc) {
@@ -987,6 +965,18 @@ export class Project implements ExternalEventEmitter {
 		} catch (e) {
 			this.logger.error('[Service#showCacheRoot]', e)
 		}
+	}
+
+	public shouldExclude(uri: string): boolean {
+		if (this.config.env.exclude.length === 0) {
+			return false
+		}
+		for (const rel of fileUtil.getRels(uri, this.projectRoots)) {
+			if (micromatch.any(rel, this.config.env.exclude)) {
+				return true
+			}
+		}
+		return false
 	}
 
 	private tryClearingCache(uri: string): void {
